% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpgpCovariates.R
\name{wpgpGetCountryCovariate}
\alias{wpgpGetCountryCovariate}
\title{wpgpListCountryCovariates function will return a list of
avalible covariates for a country}
\usage{
wpgpGetCountryCovariate(ISO3 = NULL, covariate = NULL, year = NULL,
  destDir = NULL, username = NULL, password = NULL, quiet = TRUE,
  frCSVDownload = FALSE, method = "auto")
}
\arguments{
\item{ISO3}{a country code}

\item{covariate}{Covariate name}

\item{year}{Year of the dataset you would like to download}

\item{destDir}{Path to the folder where you want to save raster file}

\item{username}{ftp username to WorldPop ftp server}

\item{password}{ftp password to WorldPop ftp server}

\item{quiet}{Download Without any messages if TRUE}

\item{frCSVDownload}{If TRUE, a new wpgAllCovariates.csv file will downloaded}

\item{method}{Method to be used for downloading files. Current download methods
are "internal", "wininet" (Windows only) "libcurl", "wget" and
"curl", and there is a value "auto"}
}
\value{
Raster file will be downloaded
}
\description{
wpgpListCountryCovariates function will return a list of
avalible covariates for a country
}
\examples{
wpgpGetCountryCovariate('NPL','px_area','2000','G:/WorldPop_Data/R_Package','ftpUsername','ftpPassword')
}
