% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpgpCovariates.R
\name{wpgpGetPOPTable}
\alias{wpgpGetPOPTable}
\title{wpgpGetPOPTable function will download a population csv}
\usage{
wpgpGetPOPTable(ISO3 = NULL, year = NULL, destDir = tempdir(),
  username = NULL, password = NULL, quiet = TRUE, overwrite = TRUE,
  method = "auto")
}
\arguments{
\item{ISO3}{a 3-character country code}

\item{year}{Year of the dataset you would like to download.}

\item{destDir}{Path to the folder where you want to save poptable file}

\item{username}{ftp username to WorldPop ftp server}

\item{password}{ftp password to WorldPop ftp server}

\item{quiet}{Download Without any messages if TRUE}

\item{overwrite}{Logical. Overwrite the poptable csv file if it already exists}

\item{method}{Method to be used for downloading files. Current download methods
are "internal", "wininet" (Windows only) "libcurl", "wget" and
"curl", and there is a value "auto"}
}
\value{
dataframe
}
\description{
wpgpGetPOPTable function will download a population csv
}
\examples{
wpgpGetPOPTable("AGO",2000,"G:/WorldPop/",username = "ftpUsername",password = "ftpPassword")
}
