% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpgpCovariates.R
\name{wpgpListCountries}
\alias{wpgpListCountries}
\title{wpgpListCountries function will return a list of the country
avalible to download}
\usage{
wpgpListCountries(username, password, verbose = FALSE,
  frCSVDownload = FALSE)
}
\arguments{
\item{username}{ftp username to WorldPop ftp server}

\item{password}{ftp password to WorldPop ftp server}

\item{verbose}{quiet If TRUE, suppress status messages (if any)}

\item{frCSVDownload}{If TRUE, a new wpgAllCovariates.csv file will downloaded}
}
\value{
Dataframe
}
\description{
wpgpListCountries function will return a list of the country
avalible to download
}
