% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpgpCovariates.R
\name{wpgpListCountryCovariates}
\alias{wpgpListCountryCovariates}
\title{wpgpListCountryCovariates function will return a list of
avalible covariates for a country}
\usage{
wpgpListCountryCovariates(ISO3 = NULL, username = NULL, password = NULL,
  detailed = FALSE)
}
\arguments{
\item{ISO3}{a country code}

\item{username}{ftp username to WorldPop ftp server}

\item{password}{ftp password to WorldPop ftp server}

\item{detailed}{If TRUE, then more information will be given}
}
\value{
Dataframe
}
\description{
wpgpListCountryCovariates function will return a list of
avalible covariates for a country
}
\examples{
wpgpListCountryCovariates( ISO3="USA", username="ftpUsername", password="ftpPassword" )
}
