% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpgpCovariates.R
\name{wpgpGetCountryCovariate}
\alias{wpgpGetCountryCovariate}
\title{wpgpGetCountryCovariate function will download files and return a list 
with the file paths to the requested covariates for one or more countries}
\usage{
wpgpGetCountryCovariate(df.user = NULL, ISO3 = NULL, covariate = NULL,
  destDir = tempdir(), username = NULL, password = NULL, quiet = TRUE,
  frCSVDownload = FALSE, method = "auto")
}
\arguments{
\item{df.user}{data frame of files to download. Must contain ISO3, Folder, and RstName.
If not supplied, must give ISO3, year, and covariate}

\item{ISO3}{a 3-character country code or vector of country codes. Optional if df.user supplied}

\item{covariate}{Covariate name(s). Optional if df.user supplied}

\item{destDir}{Path to the folder where you want to save raster file}

\item{username}{ftp username to WorldPop ftp server}

\item{password}{ftp password to WorldPop ftp server}

\item{quiet}{Download Without any messages if TRUE}

\item{frCSVDownload}{If TRUE, a new wpgAllCovariates.csv file will downloaded}

\item{method}{Method to be used for downloading files. Current download methods
are "internal", "wininet" (Windows only) "libcurl", "wget" and
"curl", and there is a value "auto"}
}
\value{
List of files downloaded, including file paths
}
\description{
wpgpGetCountryCovariate function will download files and return a list 
with the file paths to the requested covariates for one or more countries
}
\examples{
wpgpGetCountryCovariate(df.user = NULL,'NPL','px_area','G:/WorldPop/','ftpUsername','ftpPassword')
}
