% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpgpCovariates.R
\name{wpgpListCountryCovariates}
\alias{wpgpListCountryCovariates}
\title{wpgpListCountryCovariates function will return a data frame of
avalible covariates for a country}
\usage{
wpgpListCountryCovariates(ISO3 = NULL, username = NULL, password = NULL,
  detailed = FALSE, frCSVDownload = FALSE)
}
\arguments{
\item{ISO3}{a 3-character country code or vector of country codes}

\item{username}{ftp username to WorldPop ftp server}

\item{password}{ftp password to WorldPop ftp server}

\item{detailed}{If TRUE, then more information will be given}

\item{frCSVDownload}{If TRUE, a new wpgAllCovariates.csv file will downloaded}
}
\value{
Dataframe
}
\description{
wpgpListCountryCovariates function will return a data frame of
avalible covariates for a country
}
\examples{
wpgpListCountryCovariates( ISO3="USA", username="ftpUsername", password="ftpPassword" )

wpgpListCountryCovariates(ISO3=c("USA","MEX"), username="ftpUsername", password="ftpPassword" )
}
